# Faster triangle point intersection