--- a/embedded-graphics/src/primitives/triangle/mod.rs
+++ b/embedded-graphics/src/primitives/triangle/mod.rs
@@ -177,6 +177,8 @@ impl Transform for Triangle {
 }
 
 fn sort_two_yx(p1: Point, p2: Point) -> (Point, Point) {
+    // If p1.y is less than p2.y, return it first. Otherwise, if they have the same Y coordinate,
+    // the first point becomes the one with the lesser X coordinate.
     if p1.y < p2.y || (p1.y == p2.y && p1.x < p2.x) {
         (p1, p2)
     } else {
@@ -184,6 +186,8 @@ fn sort_two_yx(p1: Point, p2: Point) -> (Point, Point) {
     }
 }
 
+/// Sort 3 points in order of increasing Y value. If two points have the same Y value, the one with
+/// the lesser X value is put before.
 fn sort_yx(p1: Point, p2: Point, p3: Point) -> (Point, Point, Point) {
     let (y1, y2) = sort_two_yx(p1, p2);
     let (y1, y3) = sort_two_yx(p3, y1);