--- a/simulator/Cargo.toml
+++ b/simulator/Cargo.toml
@@ -34,6 +34,9 @@ name = "fonts"
 harness = false
 name = "image"
 required-features = ["bmp"]
+[[bench]]
+harness = false
+name = "contains"
 
 [dependencies]
 image = "0.23.0"