@jamwaffles 2020-07-13
Some benchmark results:

| Bench | Before | After | Change |
| --- | --- | --- | --- |
| triangle contains point (inside) | 111.95 ns | 7.5670 ns | -93.215% |
| triangle contains point (outside) | 2.5961 us | 118.25 ns | -95.490% |

Ryzen 7 3700X OCed to 3.8GHz, 32GB 3000MHz DDR4, Linux Mint 20

Before:

```
triangle contains point (inside)
                        time:   [111.41 ns 111.95 ns 112.53 ns]
                        change: [+0.7986% +1.1838% +1.5368%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

triangle contains point (outside)
                        time:   [2.5926 us 2.5961 us 2.6005 us]
                        change: [-0.5621% -0.2552% +0.0245%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe
```

After:

```
triangle contains point (inside)
                        time:   [7.5471 ns 7.5670 ns 7.5873 ns]
                        change: [-93.239% -93.215% -93.194%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe

triangle contains point (outside)
                        time:   [117.58 ns 118.25 ns 118.96 ns]
                        change: [-95.528% -95.490% -95.445%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```